Reduce nextTickSupply to uint128 so that the MintScheduleState structure fits in 256 bits.

+++ contracts/Lib.sol
@@ -26,5 +26,5 @@ struct MintScheduleState {
     uint8 itemIndex;
     uint16 cycleIndex;
     uint40 cycleStartTime;
-    uint256 nextTickSupply;
+    uint128 nextTickSupply;
 }
